+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.2.0] - 2024-06-15
 ### Added
 - Add `writeRecord()` to `CsvWriter` to allow writing records field by field
 - Allow overwriting the limits of 16K fields per record and 16M characters per field (#104); Thanks to [@Obolrom](https://github.com/Obolrom)!
@@ -117,7 +117,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/osiegmar/FastCSV/compare/v3.1.0...main
+[Unreleased]: https://github.com/osiegmar/FastCSV/compare/v3.2.0...main
+[3.2.0]: https://github.com/osiegmar/FastCSV/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/osiegmar/FastCSV/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/osiegmar/FastCSV/compare/v2.2.2...v3.0.0
 [2.2.2]: https://github.com/osiegmar/FastCSV/compare/v2.2.1...v2.2.2